XA: fix decoder bugs, optimizations

+++ Source/RhythmCodex.Lib/Sounds.Xa/Converters/IXaFrameSplitter.cs
@@ -5,6 +5,6 @@ namespace RhythmCodex.Sounds.Xa.Converters;
 public interface IXaFrameSplitter
 {
     int GetStatus(ReadOnlySpan<byte> frame, int channel);
-    void Get4BitData(ReadOnlySpan<byte> frame, Span<int> buffer, int channel);
-    void Get8BitData(ReadOnlySpan<byte> frame, Span<int> buffer, int channel);
+    void Get4BitData(ReadOnlySpan<byte> frame, Span<byte> buffer, int channel);
+    void Get8BitData(ReadOnlySpan<byte> frame, Span<byte> buffer, int channel);
 }
